update options type
