+++ src/types.ts
@@ -3,10 +3,20 @@ import ts from 'typescript'
 export type LiteralType = string | number | boolean
 
 export type ZodToTsOptions = {
-	/** @deprecated please use nativeEnums instead */
+	/** @deprecated use `nativeEnums` instead */
 	resolveNativeEnums?: boolean
 	nativeEnums?: 'identifier' | 'resolve' | 'union'
 }
+
+export const resolveOptions = (raw?: ZodToTsOptions) => {
+	const resolved = {
+		nativeEnums: raw?.resolveNativeEnums ? 'resolve' : 'identifier',
+	} satisfies ZodToTsOptions
+
+	return { ...resolved, ...raw }
+}
+
+export type ResolvedZodToTsOptions = ReturnType<typeof resolveOptions>
 
 export type ZodToTsStore = {
 	nativeEnums: ts.EnumDeclaration[]
@@ -20,7 +30,7 @@ export type ZodToTsReturn = {
 export type GetTypeFunction = (
 	typescript: typeof ts,
 	identifier: string,
-	options: ZodToTsOptions,
+	options: ResolvedZodToTsOptions,
 ) => ts.Identifier | ts.TypeNode
 
 export type GetType = { _def: { getType?: GetTypeFunction } }
